Sema: Move constructor synthesis code from TypeCheckDecl.cpp to CodeSynthesis.cpp
This was split between the two files; consolidate it in CodeSynthesis.cpp.
